+++ ui/src/App.tsx
@@ -57,7 +57,6 @@ import { TradingOrder } from './pages/trading/Order';
 import Notifications, { useAllNotifications } from './pages/notifications/Notifications';
 import { ServiceRequired } from './pages/error/ServiceRequired';
 import paths from './paths';
-import { partitionArray } from './pages/common';
 
 type Entry = {
   displayEntry: () => boolean;
@@ -319,17 +318,6 @@ const AppComponent = () => {
 
   const allNotifications = useAllNotifications(party);
 
-  const [notifications, pendingNotifications] = partitionArray(
-    n => n.kind != 'Pending',
-    allNotifications
-  );
-
-  const notifCount = notifications.reduce((count, { contracts }) => count + contracts.length, 0);
-  const pendingNotifCount = pendingNotifications.reduce(
-    (count, { contracts }) => count + contracts.length,
-    0
-  );
-
   entries.push({
     displayEntry: () => true,
     sidebar: [],
@@ -462,6 +450,7 @@ const AppComponent = () => {
 
   const path = useLocation().pathname;
   const currentEntry = entriesToDisplay.find(entry => path.startsWith(getBaseSegment(entry.path)));
+
   return (
     <Page
       sideBarItems={entriesToDisplay}
@@ -474,8 +463,6 @@ const AppComponent = () => {
         )
       }
       topMenuButtons={currentEntry && currentEntry.topMenuButtons}
-      showNotificationAlert={notifCount > 0}
-      showPendingNotificationAlert={pendingNotifCount > 0}
     >
       {servicesLoading ? (
         <div className="page-loading">
