+++ ui/src/pages/custody/Account.scss
@@ -64,7 +64,7 @@
       }
 
       .close-request {
-          margin-left: auto;
+        margin-left: auto;
       }
     }
   }
