+++ ui/src/pages/notifications/Notifications.scss
@@ -19,7 +19,7 @@
 
       p.pending {
         font-style: italic;
-        color: var(--cool-grey-30);
+        color: var(--yellow-40);
       }
     }
